backend/requirements.txt
fastapi==0.110.0
uvicorn==0.24.0
python-multipart==0.0.6
firebase-admin==6.4.0
numpy==1.24.3
opencv-python==4.8.1.78
pillow==10.1.0
tensorflow>=2.16.1
scikit-learn==1.3.2
requests==2.31.0
python-dotenv==1.0.0
aiofiles==23.2.1
protobuf==3.20.3
tensorflow-hub==0.15.0
tensorflow-estimator
tf-keras==2.16.0